+++ src/clj/rems/layout.clj
@@ -1,5 +1,6 @@
 (ns rems.layout
   (:require [hiccup.page :refer [html5 include-css include-js]]
+            [rems.common.git :as git]
             [rems.config :refer [env]]
             [rems.context :as context]
             [rems.db.users :as users]
@@ -23,6 +24,11 @@ window.rems = {
 (defn- css-filename [language]
   (str "/css/" (name language) "/screen.css"))
 
+(defn- cache-bust [filename]
+  (str filename "?" (or (:revision git/+version+)
+                        ;; for dev mode
+                        (System/currentTimeMillis))))
+
 (defn- page-template
   [content & [app-content]]
   (html5 [:html {:lang "en"}
@@ -35,7 +41,7 @@ window.rems = {
                      #(text :t.header/title))]
            (include-css "/assets/bootstrap/css/bootstrap.min.css")
            (include-css "/assets/font-awesome/css/all.css")
-           (include-css (css-filename (env :default-language)))]
+           (include-css (cache-bust (css-filename (env :default-language))))]
           [:body
            [:div#app app-content]
            (include-js "/assets/font-awesome/js/fontawesome.js")
@@ -64,7 +70,10 @@ window.rems = {
   [content & [params]]
   (let [content-type (:content-type params "text/html; charset=utf-8")
         status (:status params 200)
-        headers (:headers params {})]
+        ;; we don't want to cache any HTML pages since they contain
+        ;; references to cache-busted app.js and screen.css
+        headers (merge {"Cache-Control" "no-store"}
+                       (:headers params))]
     (response/content-type
      {:status status
       :headers headers
@@ -77,7 +86,7 @@ window.rems = {
     [:script {:type "text/javascript"}
      (format "var csrfToken = '%s';" (when (bound? #'*anti-forgery-token*)
                                        *anti-forgery-token*))]
-    (include-js "/js/app.js")
+    (include-js (cache-bust "/js/app.js"))
     [:script {:type "text/javascript"}
      (format "rems.app.setIdentity(%s);"
              (json/generate-string {:user (when context/*user*