+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Changes since v2.14.1
 ### Changes
 
 ### Fixes
+- Various HTTP caching issues resolved. Users should no longer get an old app.js from their browser cache. (#2484)
 
 ### Additions